boot/edk2: new package
EDK2 is a modern, feature-rich, cross-platform firmware development
environment for the UEFI and PI specifications.

The initial version of this bootloader package makes it possible to
build firmware for the following five configurations:

* QEMU x86-64 pc machine
* QEMU aarch64 virt machine, booting directly from flash
* QEMU aarch64 virt machine, booting via kernel protocol
* QEMU aarch64 sbsa-ref machine
* ARM FVP vexpress machine

When building for QEMU sbsa-ref and ARM FVP there is a dependency
on package/edk2-platforms for additional platform description files.

Signed-off-by: Dick Olsson <hi@senzilla.io>