boot/edk2: new package
EDK2 is a modern, feature-rich, cross-platform firmware development environment
for the UEFI and PI specifications.

The initial version of this bootloader package makes it possible to build flash
device images for QEMU and ARM Fixed Virtual Platforms (FVP) in different
configurations targeting AArch64.

When building for ARM FVP there is a dependency on package/edk2-platforms for
additional platform description files.

Signed-off-by: Dick Olsson <hi@senzilla.io>