configs/qemu_aarch64_sbsa_sbbr_defconfig: new config for SBBR on QEMU SBSA
This introduces a configuration for the SBSA reference machine under QEMU that
builds firmware that is close to SBBR compliance. It consists of ATF that load
EDK2 as BL33 which in turn will load the kernel as EFI stub in ACPI mode.

This patch is also introducing a generic SBSA board called aarch64_sbsa that can
be shared between SBSA platforms.

Signed-off-by: Dick Olsson <hi@senzilla.io>